Feat: Improve upload video processr to also handle trailer upload
