+++ src/lessons/processors/index.ts
@@ -1,3 +1,3 @@
-import { UploadVideoProcessor } from './upload-video/upload-video.processor';
+import { UploadVideoProcessor } from 'src/common/processors/upload-video/upload-video.processor';
 
 export const Processors = [UploadVideoProcessor];
